added a check to see if we are at correct brightnesses or not. Once at the appropriate brightness we just don't map anymore. Map seems to only map at a max or min of 32000 or -32000 which causes an overflow at around 10 mins of running.

+++ v1/lib/Dimmer/Dimmer.h
@@ -7,8 +7,8 @@ public:
   long sunriseLen; // = 600000L; // 10 minutes
   long sunsetLen;
 
-  int brightness = minBright;
-  int startBrightness = brightness;
+  unsigned long brightness = minBright;
+  unsigned long startBrightness = brightness;
   unsigned long elapsedTime = 0UL;
   unsigned long startTime = myMillis();
   unsigned long milliStart = 0UL;
